% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etwfe.R
\name{etwfe}
\alias{etwfe}
\title{Extended two-way fixed effects}
\usage{
etwfe(
  fml = NULL,
  gvar = NULL,
  gref = NULL,
  tvar = NULL,
  tref = NULL,
  data = NULL,
  cgroup = c("notyet", "never"),
  family = NULL,
  ...
)
}
\arguments{
\item{fml}{A formula with the outcome (lhs) and any additional control 
variables (rhs), e.g. `y ~ x1`. If no additional controls are required, the 
rhs must take the value of zero, e.g. `y ~ 0`.}

\item{gvar}{Character. Group variable. In staggered treatment settings this 
would correspond to a variable that denotes treatment cohorts.}

\item{gref}{Optional reference value for `gvar`. You shouldn't need to 
provide this if your `gvar` variable is well specified (by default we will
look to reference against a value greater than `max(tvar)`). But providing 
an explicit reference value can be useful/necessary if the never treated 
group, for example, takes an unusual value.}

\item{tvar}{Character. Time variable.}

\item{tref}{Optional reference value for `tvar`. Defaults to its minimum 
value (i.e., the first time period observed in the dataset).}

\item{data}{The data frame that you want to run ETWFE on.}

\item{cgroup}{Character. What control group do you wish to use for 
estimating treatment effects. Either "notyet" treated (the default) or
"never" treated.}

\item{family}{Family to be used for the estimation. Passed to 
`fixest::feglm`. Defaults to NULL, in which case `fixest::feols` is used
instead.}

\item{...}{Additional arguments passed to `fixest::feols` (or 
`fixest::feglm`). The most common example would be a `vcov` argument.}
}
\value{
A fixest object with fully saturated interaction effects.
}
\description{
Extended two-way fixed effects
}
\examples{
# We'll use the 'base_stagg' dataset from fixest to demonstrate ETWFE's
# functionality in a staggered difference-in-differences setting.
data("base_stagg", package = "fixest")

# Run the estimation
mod = etwfe(
    fml  = y ~ x1, 
    gvar = "year_treated", 
    tvar = "year", 
    data = base_stagg, 
    vcov = ~ id
    )
mod

# We can recover a variety of treatment effects of interest with the 
# complementary emfx() function. For example:
emfx(mod, type = "event")

}
