% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.emfx}
\alias{plot.emfx}
\title{Plot method for emfx objects}
\usage{
\method{plot}{emfx}(
  x,
  type = c("pointrange", "errorbar", "ribbon"),
  pch = 16,
  zero = TRUE,
  grid = TRUE,
  ref = -1,
  ...
)
}
\arguments{
\item{x}{An \code{emfx} object.}

\item{type}{Character. The type of plot display. One of \code{"pointrange"}
(default), \code{"errorbar"}, or \code{"ribbon"}.}

\item{pch}{Integer or character. Which plotting character or symbol to use
(see \code{\link[graphics]{points}}). Defaults to 16 (i.e., small solid
circle). Ignored if \code{type = "ribbon"}.}

\item{zero}{Logical. Should 0-zero line be emphasized? Default is \code{TRUE}.}

\item{grid}{Logical. Should a background grid be displayed? Default is
\code{TRUE}.}

\item{ref}{Integer. Reference line marker for event-study plot. Default is
\code{-1} (i.e., the period immediately preceding treatment). To remove
completely, set to any one of \code{NA}, \code{NULL}, or \code{FALSE}. Only used if the
underlying object was computed using \code{emfx(..., type = "event")}.}

\item{...}{Additional arguments passed to \code{\link[tinyplot:tinyplot]{tinyplot::tinyplot}}.}
}
\value{
No return value, called for side effect of producing a plot.
}
\description{
Plot method for emfx objects
}
