% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emfx.R
\name{emfx}
\alias{emfx}
\title{Post-estimation treatment effects for an ETWFE regressions.}
\usage{
emfx(
  object,
  type = c("simple", "group", "calendar", "event"),
  by_xvar = "auto",
  collapse = "auto",
  post_only = TRUE,
  ...
)
}
\arguments{
\item{object}{An `etwfe` model object.}

\item{type}{Character. The desired type of post-estimation aggregation.}

\item{by_xvar}{Logical. Should the results account for heterogeneous
treatment effects? Only relevant if the preceding `etwfe` call included a
specified `xvar` argument, i.e. interacted categorical covariate. The
default behaviour ("auto") is to automatically estimate heterogeneous
treatment effects for each level of `xvar` if these are detected as part
of the underlying `etwfe` model object. Users can override by setting to
either FALSE or TRUE. See the section on Heterogeneous treatment effects
below.}

\item{collapse}{Logical. Collapse the data by (period by cohort) groups
before calculating marginal effects? This trades off a loss in estimate
accuracy (typically around the 1st or 2nd significant decimal point) for a
substantial improvement in estimation time for large datasets. The default
behaviour ("auto") is to automatically collapse if the original dataset
has more than 500,000 rows. Users can override by setting either FALSE or
TRUE. Note that collapsing by group is only valid if the preceding `etwfe`
call was run with "ivar = NULL" (the default). See the section on
Performance tips below.}

\item{post_only}{Logical. Only keep post-treatment effects. All
pre-treatment effects will be zero as a mechanical result of ETWFE's
estimation setup, so the default is to drop these nuisance rows from the
dataset. But you may want to keep them for presentation reasons (e.g.,
plotting an event-study); though be warned that this is strictly
performative. This argument will only be evaluated if `type = "event"`.}

\item{...}{Additional arguments passed to
[`marginaleffects::marginaleffects`]. For example, you can pass `vcov =
FALSE` to dramatically speed up estimation times of the main marginal
effects (but at the cost of not getting any information about standard
errors; see Performance tips below). Another potentially useful
application is testing whether heterogeneous treatment effects (i.e. the
levels of any `xvar` covariate) are equal by invoking the `hypothesis`
argument, e.g. `hypothesis = "b1 = b2"`.}
}
\value{
A `slopes` object from the `marginaleffects` package.
}
\description{
Post-estimation treatment effects for an ETWFE regressions.
}
\section{Performance tips}{
 

  For datasets smaller than 100k rows, `emfx` should complete quite
  quickly; within a few seconds or less. However, the computation time does
  tend to scale linearly with the size of the data, as well as the number of
  interactions from the original `etwfe` model. Without getting too far into
  the weeds, the delta method of the underlying marginal effects calculation
  has to estimate two prediction models for *each* coefficient in the model
  and then compute their standard errors. So, it's a potentially expensive
  operation. 
  
  However, there are two key strategies that you can use to speed things up.
  The first is to pass "vcov = FALSE" as an argument to `emfx`. Doing so
  should reduce the estimation time to less than a second, even for datasets
  in excess of a million rows. This approach does come at the cost of not
  returning any standard errors. Yet it can be useful to combine our first
  strategy with a second strategy, which is to invoke the "collapse = TRUE"
  argument. Collapsing the data by groups prior to estimating the marginal
  effects can yield a substantial speed increase (albeit not nearly as
  dramatic as turning of the vcov calculations). But we do get standard
  errors this time. The trade-off from collapsing the data is that we lose
  some accuracy in our estimated parameters. Testing suggests that this loss
  in accuracy tends to be relatively minor, with results equivalent to the
  1st or 2nd significant decimal place (or even better). By combining these
  two strategies, users can very quickly see how bad the loss in accuracy is
  on the main marginal effects, before deciding whether to estimate with the
  collapsed dataset to get approximate standard errors.
  
  Summarizing, if you are worried about the estimation time for a large
  dataset, try the following three-step approach:
  
  1. Run `emfx(..., vcov = FALSE)`.
  
  2. Run `emfx(..., vcov = FALSE, collapse = TRUE)`.
  
  3. Compare the results from steps 1 and 2. If the main parameter estimates
  are similar enough, then as your final model run the following to also 
  obtain approximate standard errors: `emfx(..., collapse = TRUE)`.
}

\section{Heterogeneous treatment effects}{


  Specifying `etwfe(..., xvar = <xvar>)` will generate interaction effects
  for all levels of `<xvar>` as part of the main regression model. The
  reason that this is useful (as opposed to a regular, non-interacted
  covariate in the formula RHS) is that it allows us to estimate
  heterogeneous treatment effects as part of the larger ETWFE framework.
  Specifically, we can recover heterogeneous treatment effects for each
  level of `<xvar>` by passing the resulting `etwfe` model object on to 
  `emfx()`.
  
  For example, imagine that we have a categorical variable called "age" in
  our dataset, with two distinct levels "adult" and "child". Running
  `emfx(etwfe(..., xvar = age))` will tell us how the efficacy of treatment 
  varies across adults and children. We can then also leverage the in-built 
  hypothesis testing infrastructure of `marginaleffects` to test whether
  the treatment effect is statistically different across these two age
  groups; see Examples below. Note the same principles carry over to 
  categorical variables with multiple levels, or even continuous variables
  (although continuous variables are not as well supported yet).
}

\examples{
# We’ll use the mpdta dataset from the did package (which you’ll need to 
# install separately).

# install.packages("did")
data("mpdta", package = "did")

#
# Basic example
#

# The basic ETWFE workflow involves two steps:

# 1) Estimate the main regression model with etwfe().

mod = etwfe(
    fml  = lemp ~ lpop, # outcome ~ controls (use 0 or 1 if none)
    tvar = year,        # time variable
    gvar = first.treat, # group variable
    data = mpdta,       # dataset
    vcov = ~countyreal  # vcov adjustment (here: clustered by county)
    )
mod

# 2) Recover the treatment effects of interest with emfx().

emfx(mod)                 # simple average treatment effect (default)
emfx(mod, type = "event") # dynamic treatment effect a la an event study
# Etc. Other aggregation types are "group" and "calendar"


#
# Heterogeneous treatment effects
#

# Example where we estimate heterogeneous treatment effects for counties 
# within the 8 US Great Lake states (versus all other counties). 

gls = c("IL" = 17, "IN" = 18, "MI" = 26, "MN" = 27,
        "NY" = 36, "OH" = 39, "PA" = 42, "WI" = 55)

mpdta$gls = substr(mpdta$countyreal, 1, 2) \%in\% gls

hmod = etwfe(
   lemp ~ lpop, tvar = year, gvar = first.treat, data = mpdta, 
   vcov = ~countyreal,
   xvar = gls           ## <= het. TEs by gls
   )

# Heterogeneous ATEs (could also specify "event", etc.) 

emfx(hmod)

# To test whether the ATEs across these two groups (non-GLS vs GLS) are 
# statistically different, simply pass an appropriate "hypothesis" argument.

emfx(hmod, hypothesis = "b1 = b2")


#
# Nonlinear model (distribution / link) families
#

# Poisson example

mpdta$emp = exp(mpdta$lemp)

etwfe(
   emp ~ lpop, tvar = year, gvar = first.treat, data = mpdta, 
   vcov = ~countyreal,
   family = "poisson"   ## <= family arg for nonlinear options
   ) |>
   emfx("event")

}
\seealso{
[marginaleffects::slopes()]
}
