% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emfx.R
\name{emfx}
\alias{emfx}
\title{Post-estimation treatment effects for an ETWFE regressions.}
\usage{
emfx(object, type = c("simple", "group", "calendar", "event"), ...)
}
\arguments{
\item{object}{An `etwfe` model object.}

\item{type}{Character. The desired type of post-estimation aggregation.}

\item{...}{Additional arguments passed to 
`marginaleffects::marginaleffects`.}
}
\value{
A marginaleffects object.
}
\description{
Post-estimation treatment effects for an ETWFE regressions.
}
