% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emfx.R
\name{emfx}
\alias{emfx}
\title{Post-estimation treatment effects for an ETWFE regressions.}
\usage{
emfx(
  object,
  type = c("simple", "group", "calendar", "event"),
  xvar = NULL,
  post_only = TRUE,
  collapse = "auto",
  ...
)
}
\arguments{
\item{object}{An `etwfe` model object.}

\item{type}{Character. The desired type of post-estimation aggregation.}

\item{xvar}{Optional interacted categorical covariate for estimating
heterogeneous treatment effects. In other words, allows recovery of the
(marginal) treatment effect for distinct values of `xvar`. Works with binary
categorical variables (e.g. "adult" or "child"), as well as multiple values.}

\item{post_only}{Logical. Only keep post-treatment effects. All
pre-treatment effects will be zero as a mechanical result of ETWFE's 
estimation setup, so the default is to drop these nuisance rows from
the dataset. But you may want to keep them for presentation reasons
(e.g., plotting an event-study); though be warned that this is 
strictly performative. This argument will only be evaluated if
`type = "event"`.}

\item{collapse}{Logical. Collapse the data by (period by cohort) groups 
before calculating marginal effects? This trades off a slight loss in
estimate accuracy (typically around the second decimal point) for a 
substantial improvement in estimation time for large datasets. The default 
behaviour ("auto") is to automatically collapse if the original dataset has 
more than 100,000 rows. Users can override by setting either FALSE or TRUE. 
Note that collapsing by group is only valid if the preceding `etwfe` call 
was run with "ivar = NULL" (the default).}

\item{...}{Additional arguments passed to [`marginaleffects::marginaleffects`]. 
A potentially useful case is testing whether heterogeneous treatment effects
(from any `xvar` covariate) are equal by invoking the `hypothesis` argument,
e.g. `hypothesis = "adult = child"`.}
}
\value{
A `slopes` object from the `marginaleffects` package.
}
\description{
Post-estimation treatment effects for an ETWFE regressions.
}
\examples{
# We’ll use the mpdta dataset from the did package (which you’ll need to 
# install separately).

# install.packages("did")
data("mpdta", package = "did")

# Run the estimation
mod = etwfe(
    fml  = lemp ~ lpop, # outcome ~ controls
    tvar = year,        # time variable
    gvar = first.treat, # group variable
    data = mpdta,       # dataset
    vcov = ~countyreal  # vcov adjustment (here: clustered by county)
    )
mod

# We can recover a variety of treatment effects of interest with the 
# complementary emfx() function. For example:
emfx(mod, type = "event")

}
\seealso{
[marginaleffects::slopes()]
}
