% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etwfe.R
\name{etwfe}
\alias{etwfe}
\title{Extended two-way fixed effects}
\usage{
etwfe(
  fml = NULL,
  tvar = NULL,
  gvar = NULL,
  data = NULL,
  ivar = NULL,
  xvar = NULL,
  tref = NULL,
  gref = NULL,
  cgroup = c("notyet", "never"),
  fe = c("vs", "feo", "none"),
  family = NULL,
  ...
)
}
\arguments{
\item{fml}{A formula with the outcome (lhs) and any time-constant controls 
variables (rhs), e.g. `y ~ x1 + x2`. Please note that time-varying controls
are not supported. Similarly, if no additional controls are required, the 
rhs must take the value of 0 or 1, e.g. `y ~ 0`.}

\item{tvar}{Time variable. Can be a string (e.g., "year") or an expression
(e.g., year).}

\item{gvar}{Group variable. Can be either a string (e.g., "first_treated") 
or an expression (e.g., first_treated). In a staggered treatment setting, 
the group variable typically denotes treatment cohort.}

\item{data}{The data frame that you want to run ETWFE on.}

\item{ivar}{Optional index variable. Can be a string (e.g., "country") or an 
expression (e.g., country). Leaving as NULL (the default) will result in
group-level fixed effects being used, which is more efficient and necessary 
for nonlinear models (see `family` argument below). However, you may still
want to cluster your standard errors by your index variable through the
`vcov` argument. See examples below.}

\item{xvar}{Optional interacted categorical covariate for estimating
heterogeneous treatment effects. Enables recovery of the marginal
treatment effect for distinct levels of `xvar`, e.g. "child", "teenager",
or "adult". (Note that the "x" part of "xvar" is meant to represent a
treatment-*interacted* covariate, as opposed to a regular control
variable.)}

\item{tref}{Optional reference value for `tvar`. Defaults to its minimum 
value (i.e., the first time period observed in the dataset).}

\item{gref}{Optional reference value for `gvar`. You shouldn't need to 
provide this if your `gvar` variable is well specified. But providing an 
explicit reference value can be useful/necessary if the desired control 
group takes an unusual value.}

\item{cgroup}{What control group do you wish to use for 
estimating treatment effects. Either "notyet" treated (the default) or
"never" treated.}

\item{fe}{What level of fixed effects should be used? Defaults to "vs" 
(varying slopes), which is the most efficient in terms of estimation and 
terseness of the return model object. The other two options, "feo" (fixed 
effects only) and "none" (no fixed effects whatsoever), trade off efficiency
for additional information on other (nuisance) model parameters. Note that
the primary treatment parameters of interest should remain unchanged 
regardless of choice.}

\item{family}{Which [`family`] to use for the estimation. Defaults to NULL, in 
which case [`fixest::feols`] is used. Otherwise passed to [`fixest::feglm`], so
that valid entries include "logit", "poisson", and "negbin". Note that if a
non-NULL family entry is detected, `ivar` will automatically be set to NULL.}

\item{...}{Additional arguments passed to [`fixest::feols`] (or 
[`fixest::feglm`]). The most common example would be a `vcov` argument.}
}
\value{
A fixest object with fully saturated interaction effects.

A `slopes` object from the `marginaleffects` package.
}
\description{
Extended two-way fixed effects
}
\section{Heterogeneous treatment effects}{


  Specifying `etwfe(..., xvar = <xvar>)` will generate interaction effects
  for all levels of `<xvar>` as part of the main regression model. The
  reason that this is useful (as opposed to a regular, non-interacted
  covariate in the formula RHS) is that it allows us to estimate
  heterogeneous treatment effects as part of the larger ETWFE framework.
  Specifically, we can recover heterogeneous treatment effects for each
  level of `<xvar>` by passing the resulting `etwfe` model object on to 
  `emfx()`.
  
  For example, imagine that we have a categorical variable called "age" in
  our dataset, with two distinct levels "adult" and "child". Running
  `emfx(etwfe(..., xvar = age))` will tell us how the efficacy of treatment 
  varies across adults and children. We can then also leverage the in-built 
  hypothesis testing infrastructure of `marginaleffects` to test whether
  the treatment effect is statistically different across these two age
  groups; see Examples below. Note the same principles carry over to 
  categorical variables with multiple levels, or even continuous variables
  (although continuous variables are not as well supported yet).
}

\section{Performance tips}{
 

  For datasets smaller than 100k rows, `emfx` should complete quite
  quickly; within a few seconds or less. However, the computation time does
  tend to scale linearly with the size of the data, as well as the number of
  interactions from the original `etwfe` model. Without getting too far into
  the weeds, the delta method of the underlying marginal effects calculation
  has to estimate two prediction models for *each* coefficient in the model
  and then compute their standard errors. So, it's a potentially expensive
  operation. 
  
  However, there are two key strategies that you can use to speed things up.
  The first is to pass "vcov = FALSE" as an argument to `emfx`. Doing so
  should reduce the estimation time to less than a second, even for datasets
  in excess of a million rows. This approach does come at the cost of not
  returning any standard errors. Yet it can be useful to combine our first
  strategy with a second strategy, which is to invoke the "collapse = TRUE"
  argument. Collapsing the data by groups prior to estimating the marginal
  effects can yield a substantial speed increase (albeit not nearly as
  dramatic as turning of the vcov calculations). But we do get standard
  errors this time. The trade-off from collapsing the data is that we lose
  some accuracy in our estimated parameters. Testing suggests that this loss
  in accuracy tends to be relatively minor, with results equivalent to the
  1st or 2nd significant decimal place (or even better). By combining these
  two strategies, users can very quickly see how bad the loss in accuracy is
  on the main marginal effects, before deciding whether to estimate with the
  collapsed dataset to get approximate standard errors.
  
  Summarizing, if you are worried about the estimation time for a large
  dataset, try the following three-step approach:
  
  1. Run `emfx(..., vcov = FALSE)`.
  
  2. Run `emfx(..., vcov = FALSE, collapse = TRUE)`.
  
  3. Compare the results from steps 1 and 2. If the main parameter estimates
  are similar enough, then as your final model run the following to also 
  obtain approximate standard errors: `emfx(..., collapse = TRUE)`.
}

\examples{
# We’ll use the mpdta dataset from the did package (which you’ll need to 
# install separately).

# install.packages("did")
data("mpdta", package = "did")

#
# Basic example
#

# The basic ETWFE workflow involves two steps:

# 1) Estimate the main regression model with etwfe().

mod = etwfe(
    fml  = lemp ~ lpop, # outcome ~ controls (use 0 or 1 if none)
    tvar = year,        # time variable
    gvar = first.treat, # group variable
    data = mpdta,       # dataset
    vcov = ~countyreal  # vcov adjustment (here: clustered by county)
    )
mod

# 2) Recover the treatment effects of interest with emfx().

emfx(mod)                 # simple average treatment effect (default)
emfx(mod, type = "event") # dynamic treatment effect a la an event study
# Etc. Other aggregation types are "group" and "calendar"


#
# Heterogeneous treatment effects
#

# Example where we estimate heterogeneous treatment effects for counties 
# within the 8 US Great Lake states (versus all other counties). 

gls = c("IL" = 17, "IN" = 18, "MI" = 26, "MN" = 27,
        "NY" = 36, "OH" = 39, "PA" = 42, "WI" = 55)

mpdta$gls = substr(mpdta$countyreal, 1, 2) \%in\% gls

hmod = etwfe(
   lemp ~ lpop, tvar = year, gvar = first.treat, data = mpdta, 
   vcov = ~countyreal,
   xvar = gls           ## <= het. TEs by gls
   )

# Heterogeneous ATEs (could also specify "event", etc.) 

emfx(hmod)

# To test whether the ATEs across these two groups (non-GLS vs GLS) are 
# statistically different, simply pass an appropriate "hypothesis" argument.

emfx(hmod, hypothesis = "b1 = b2")


#
# Nonlinear model (distribution / link) families
#

# Poisson example

mpdta$emp = exp(mpdta$lemp)

etwfe(
   emp ~ lpop, tvar = year, gvar = first.treat, data = mpdta, 
   vcov = ~countyreal,
   family = "poisson"   ## <= family arg for nonlinear options
   ) |>
   emfx("event")

}
\references{
Wooldridge, Jeffrey M. (2021). \cite{Two-Way Fixed Effects, the 
Two-Way Mundlak Regression, and Difference-in-Differences Estimators}.
Working paper (version: August 16, 2021). Available: 
http://dx.doi.org/10.2139/ssrn.3906345

Wooldridge, Jeffrey M. (2022). \cite{Simple Approaches to Nonlinear Difference-in-Differences with Panel Data}.
Working paper (version: August 7, 2022). Available: 
http://dx.doi.org/10.2139/ssrn.4183726
}
\seealso{
[fixest::feols()], [fixest::feglm()]
}
